The various Microsoft nuget package version 7 do not support netcoreapp therefore I am removing netcoreapp3.1 from unit testing frameworks
Here's an example of the warning you get when you use one of these packages in netcoreapp3.1:
Microsoft.Extensions.DependencyInjection.Abstractions 7.0.0 doesn't support netcoreapp3.1 and has not been tested with it. Consider upgrading your TargetFramework to net6.0 or later. You may also set <SuppressTfmSupportBuildWarnings>true</SuppressTfmSupportBuildWarnings> in the project file to ignore this warning and attempt to run in this unsupported configuration at your own risk.